♻️ refactor(webview/settings): 重构设置视图，分离关注点并改进配置管理
- 【重构】将 `SettingsViewProvider.ts` 拆分为独立的 `SettingsViewHTMLProvider.ts` 和 `SettingsViewMessageHandler.ts` 以实现关注点分离
- 【重构】`SettingsViewProvider` 现在使用 `EmbeddingServiceManager` 获取 `EmbeddingService` 实例，并传递 `extensionContext` 给消息处理器
- 【新增】`SettingsViewMessageHandler.ts` 用于处理来自 webview 的所有消息：
    - 管理设置的加载，现在会合并来自 `package.json` (VS Code 配置) 和工作区状态 (`stateManager` 与 `WORKSPACE_CONFIG_SCHEMA`) 的设置项
    - 管理设置的保存，根据设置项来源 (`fromPackageJSON` 标志) 将其保存到 VS Code 配置或工作区状态
    - 处理索引启动 (`startIndexing`)，并对 `EmbeddingService` 实例为空的情况进行检查
    - 处理获取模型列表 (`getModelsForProvider`)
- 【新增】`SettingsViewHTMLProvider.ts` 专注于生成设置页面的 HTML 内容
- 【重构】`settings-page.tsx` (UI):
    - 引入新的 `SettingsMenu` 和 `SettingsContent` 子组件，以更好地组织 UI 结构
    - 使用新的 `useMessageHandler` 自定义 hook 来封装与 VS Code 扩展后端的消息通信和状态管理逻辑
    - 更新了嵌入提供商 (`embeddingProvider`) 和模型 (`embeddingModels`) 的选择逻辑，以支持动态模型列表
- 【构建】更新 `tsconfig.app.tsbuildinfo` 以反映项目文件结构的更改
